refactor: 헤더 관련 로직 useHeader 훅에 분리
