+++ src/hooks/index.ts
@@ -6,6 +6,7 @@ import useWrite from './useWrite';
 import useEdtiBoard from './useEditBoard';
 import useProfileImage from './useProfileImage';
 import useUpdateProfile from './useUpdateProfile';
+import useHeader from './useHeader';
 
 export {
   useAuth,
@@ -16,4 +17,5 @@ export {
   useEdtiBoard,
   useProfileImage,
   useUpdateProfile,
+  useHeader,
 };
